pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "capx"
version = "0.1.2"
description = "reCAPTCHA v2 Image Solver using YOLO and Selenium"
authors = [{name = "Mahdi Marjani", email = "mahdi.marjani.dev@gmail.com"}]
license = "MIT"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "numpy==2.2.6",
    "opencv-python==4.12.0.88",
    "Pillow==12.0.0",
    "requests==2.32.5",
    "selenium==4.39.0",
    "ultralytics==8.3.241",
    "webdriver-manager==4.0.2",
    "tqdm==4.67.1",
]

[tool.setuptools]
packages = ["capx"]
package-dir = {"" = "src"}

[project.urls]
Homepage = "https://github.com/mahdi-marjani/recaptcha-bypass"
Documentation = "https://github.com/mahdi-marjani/recaptcha-bypass"
Repository = "https://github.com/mahdi-marjani/recaptcha-bypass.git"
Issues = "https://github.com/mahdi-marjani/recaptcha-bypass/issues"